Support nocomments=false
I use netlify cms to handle my hugo blog, and when you define a front matter param, it always needs to set it, be it true or false.

I think that the mere existence of the param is not the best way to test this. If you set `nocomments: false` in your front matter, IMHO it should display comments.